Backend/requirements.txt
chromadb==0.5.3
fastapi
fastembed
firebase-admin==6.6.0
joblib
langchain
langchain-anthropic
langchain-community
langchain-google-genai
langchain-chroma
langchain-groq
langchain-text-splitters
llama_index
llama_parse==0.5.19
openpyxl
pdf2image
pdfservices-sdk
pypdf
PyPDF2
pytest
python-dotenv
python-multipart==0.0.20
pytesseract
requests
uvicorn
easyocr 
pdf2image
langchain_chroma